requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
motor==2.5.1
pydantic==1.9.0
PyJWT==2.1.0
bcrypt==3.2.0
openai==0.27.0
email-validator==1.1.3
python-multipart==0.0.5
pytest==6.2.5
pytest-asyncio==0.15.1
httpx==0.18.2
asgi-lifespan==1.0.1
groq
